# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zakrely API",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/subjects": {
      "get": {
        "tags": [
          "subjects"
        ],
        "summary": "List Subjects",
        "operationId": "list_subjects_subjects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubjectOut"
                  },
                  "type": "array",
                  "title": "Response List Subjects Subjects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/subjects/{slug}/path": {
      "get": {
        "tags": [
          "subjects"
        ],
        "summary": "Get Path",
        "operationId": "get_path_subjects__slug__path_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SubjectSlugEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PathNode"
                  },
                  "title": "Response Get Path Subjects  Slug  Path Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson_id}": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "Get Lesson",
        "operationId": "get_lesson_lessons__lesson_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson_id}/quiz": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "Get Quiz",
        "operationId": "get_quiz_lessons__lesson_id__quiz_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/practice/next": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Practice Next",
        "operationId": "practice_next_practice_next_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SubjectSlugEnum"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/practice/submit": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Practice Submit",
        "operationId": "practice_submit_practice_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/chat/sessions": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Session",
        "operationId": "create_session_chat_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/chat/sessions/{session_id}/message": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "operationId": "send_message_chat_sessions__session_id__message_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatMessageRequest"
      },
      "ChatMessageResponse": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "done": {
            "type": "boolean",
            "title": "Done",
            "default": false
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "reply"
        ],
        "title": "ChatMessageResponse"
      },
      "ChatModeEnum": {
        "type": "string",
        "enum": [
          "english_convo",
          "science_explain"
        ],
        "title": "ChatModeEnum"
      },
      "ChatSessionCreateRequest": {
        "properties": {
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "mode": {
            "$ref": "#/components/schemas/ChatModeEnum"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "mode"
        ],
        "title": "ChatSessionCreateRequest"
      },
      "ChatSessionCreateResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "ChatSessionCreateResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LessonDetail": {
        "properties": {
          "lesson": {
            "$ref": "#/components/schemas/LessonOut"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/LessonSectionOut"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "lesson",
          "sections"
        ],
        "title": "LessonDetail"
      },
      "LessonOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_id",
          "order_index",
          "title",
          "objective"
        ],
        "title": "LessonOut"
      },
      "LessonSectionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "heading": {
            "type": "string",
            "title": "Heading"
          },
          "body_md": {
            "type": "string",
            "title": "Body Md"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_index",
          "heading",
          "body_md"
        ],
        "title": "LessonSectionOut"
      },
      "LessonStatusEnum": {
        "type": "string",
        "enum": [
          "locked",
          "unlocked",
          "completed"
        ],
        "title": "LessonStatusEnum"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "PathNode": {
        "properties": {
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/LessonStatusEnum"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "order",
          "title",
          "status"
        ],
        "title": "PathNode"
      },
      "PracticeSetOut": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionPublic"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "questions"
        ],
        "title": "PracticeSetOut"
      },
      "PracticeSubmitRequest": {
        "properties": {
          "answers": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "PracticeSubmitRequest"
      },
      "PracticeSubmitResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/QuestionResult"
            },
            "type": "array",
            "title": "Results"
          },
          "skill_breakdown": {
            "items": {
              "$ref": "#/components/schemas/SkillBreakdownItem"
            },
            "type": "array",
            "title": "Skill Breakdown"
          }
        },
        "type": "object",
        "required": [
          "results",
          "skill_breakdown"
        ],
        "title": "PracticeSubmitResponse"
      },
      "QuestionPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "skill_tag": {
            "$ref": "#/components/schemas/SkillTagEnum"
          },
          "qtype": {
            "$ref": "#/components/schemas/QuestionTypeEnum"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "options": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lesson_id",
          "skill_tag",
          "qtype",
          "body"
        ],
        "title": "QuestionPublic",
        "description": "Question as served to a student \u2014 correct_answer is never included."
      },
      "QuestionResult": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "given_answer": {
            "type": "string",
            "title": "Given Answer"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "skill_tag": {
            "$ref": "#/components/schemas/SkillTagEnum"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "given_answer",
          "is_correct",
          "correct_answer",
          "explanation",
          "skill_tag"
        ],
        "title": "QuestionResult"
      },
      "QuestionTypeEnum": {
        "type": "string",
        "enum": [
          "mcq",
          "short_answer",
          "numeric"
        ],
        "title": "QuestionTypeEnum"
      },
      "QuizOut": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionPublic"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "questions"
        ],
        "title": "QuizOut"
      },
...
      "RegisterRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "role": {
            "$ref": "#/components/schemas/UserRoleEnum",
            "default": "student"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
      "SkillBreakdownItem": {
        "properties": {
          "skill_tag": {
            "$ref": "#/components/schemas/SkillTagEnum"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "insufficient_data": {
            "type": "boolean",
            "title": "Insufficient Data",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "skill_tag",
          "correct",
          "total"
        ],
        "title": "SkillBreakdownItem"
      },
      "SkillTagEnum": {
        "type": "string",
        "enum": [
          "memorization",
          "comprehension",
          "application",
          "analysis"
        ],
        "title": "SkillTagEnum"
      },
      "SubjectOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "$ref": "#/components/schemas/SubjectSlugEnum"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name_ar",
          "name_en"
        ],
        "title": "SubjectOut"
      },
      "SubjectSlugEnum": {
        "type": "string",
        "enum": [
          "english",
          "math",
          "science"
        ],
        "title": "SubjectSlugEnum"
      },
      "TokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "title": "TokenResponse"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRoleEnum"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role"
        ],
        "title": "UserOut"
      },
      "UserRoleEnum": {
        "type": "string",
        "enum": [
          "student",
          "parent"
        ],
        "title": "UserRoleEnum"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}